+++ primitives/src/lib.rs
@@ -19,6 +19,6 @@
 #![warn(missing_docs)]
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod v0;
+//pub mod v0;
 pub mod v1;
 pub mod v2;
